+++ jwst/extract_1d/extract.py
@@ -8,9 +8,9 @@ import numpy as np
 from astropy.modeling import polynomial
 from .. import datamodels
 from ..datamodels import dqflags
-from .. assign_wcs import niriss        # for specifying spectral order number
-from .. transforms import models as trmodels
-from .. lib import pipe_utils
+from ..assign_wcs import niriss         # for specifying spectral order number
+from ..assign_wcs.util import wcs_bbox_from_shape
+from ..lib import pipe_utils
 from . import extract1d
 from . import ifu
 from . import spec_wcs
@@ -58,6 +58,10 @@ HORIZONTAL = 1
 VERTICAL = 2
 """Dispersion direction, predominantly horizontal or vertical."""
 
+# This is intended to be larger than any possible distance (in pixels)
+# between the target and any point in the image; used by locn_from_wcs().
+HUGE_DIST = 1.e20
+
 # These values are assigned in get_extract_parameters, using key "match".
 # If there was an aperture in the reference file for which the "id" key
 # matched, that's (at least) a partial match.  If "spectral_order" also
@@ -66,20 +70,6 @@ VERTICAL = 2
 NO_MATCH = "no match"
 PARTIAL = "partial match"
 EXACT = "exact match"
-
-DUMMY = "dummy"
-"""Distinguish between a slit object and not a slit object.
-
-Extended summary
-----------------
-The input model can contain a sequence of slits (SlitModel objects,
-2-D cutouts, with attributes), or it might be a full detector image or
-subarray.  In the former case, the step will try to extract a spectrum
-from each slit.  In the latter case, the input can be a CubeModel or an
-ImageModel, and not all of the attributes are the same.  DUMMY is
-passed to functions instead of a SlitModel object if the input model
-does not contain slits.
-"""
 
 Aperture = namedtuple('Aperture', ['xstart', 'ystart', 'xstop', 'ystop'])
 
@@ -143,7 +133,8 @@ def load_ref_file(refname):
 
 def get_extract_parameters(ref_dict,
                            input_model, slitname, sp_order,
-                           meta, smoothing_length, bkg_order):
+                           meta, smoothing_length, bkg_order,
+                           apply_nod_offset):
     """Get reference file values.
 
     Parameters
@@ -191,13 +182,20 @@ def get_extract_parameters(ref_dict,
         This argument must be positive or zero, and it is only used if
         background regions have been specified.
 
+    apply_nod_offset : bool or None
+        If True, the target and background positions specified in `ref_dict`
+        (or a default target position) will be shifted to account for nod
+        and/or dither offset.
+        If None, the value specified in `ref_dict` will be used, or it will
+        be set to True if not found in `ref_dict`.
+
     Returns
     -------
     extract_params : dict
         Information copied out of `ref_dict`.  The items will be selected
         based on `slitname` and `sp_order`.  Default values will be
         assigned if `ref_dict` is None.  For a reference image, the key
-        'ref_image' gives the image model.
+        'ref_image' gives the (open) image model.
     """
 
     extract_params = {'match': NO_MATCH}        # initial value
@@ -215,6 +213,10 @@ def get_extract_parameters(ref_dict,
         extract_params['extract_width'] = None
         extract_params['src_coeff'] = None
         extract_params['bkg_coeff'] = None
+        if apply_nod_offset is None:
+            extract_params['apply_nod_offset'] = True
+        else:
+            extract_params['apply_nod_offset'] = apply_nod_offset
         extract_params['nod_correction'] = 0
         extract_params['independent_var'] = 'pixel'
         extract_params['smoothing_length'] = 0  # because no background sub.
@@ -267,6 +269,12 @@ def get_extract_parameters(ref_dict,
                     else:
                         # If the user supplied a value, use that value.
                         extract_params['bkg_order'] = bkg_order
+                    if apply_nod_offset is None:
+                        extract_params['apply_nod_offset'] = \
+                              aper.get('apply_nod_offset', True)
+                    else:
+                        # If the user supplied a value, use that value.
+                        extract_params['apply_nod_offset'] = apply_nod_offset
                     extract_params['xstart'] = aper.get('xstart')
                     extract_params['xstop'] = aper.get('xstop')
                     extract_params['ystart'] = aper.get('ystart')
@@ -300,6 +308,10 @@ def get_extract_parameters(ref_dict,
             else:
                 # The user-supplied value takes precedence.
                 extract_params['smoothing_length'] = smoothing_length
+            if apply_nod_offset is None:
+                extract_params['apply_nod_offset'] = True
+            else:
+                extract_params['apply_nod_offset'] = apply_nod_offset
             extract_params['nod_correction'] = 0
 
     else:
@@ -317,8 +329,8 @@ def find_dispaxis(input_model, slit, spectral_order, extract_params):
     input_model : data model
         The input science file.
 
-    slit : data model or str
-        This is a slit from a `MultiSlitModel` (or similar), or "dummy"
+    slit : data model or None
+        This is a slit from a `MultiSlitModel` (or similar), or None
         if the input is not an array of 2-D cutouts.
         We use the `meta.wcs` and `wavelength` attributes.
 
@@ -345,13 +357,13 @@ def find_dispaxis(input_model, slit, spectral_order, extract_params):
         # updated with a valid value, we will not extract the spectrum.
         extract_params['dispaxis'] = None
 
-    if slit == DUMMY:
+    if slit is None:
         shape = input_model.data.shape[-2:]
     else:
         shape = slit.data.shape[-2:]
 
     wcs = None                                  # initial value
-    if slit == DUMMY:
+    if slit is None:
         if input_model.meta.exposure.type == "NIS_SOSS":
             if hasattr(input_model.meta, 'wcs'):
                 try:
@@ -396,7 +408,7 @@ def find_dispaxis(input_model, slit, spectral_order, extract_params):
         log.warning("Can't determine dispaxis from the WCS.")
         return
 
-    if slit == DUMMY:
+    if slit is None:
         got_wavelength = False
     else:
         if hasattr(slit, "wavelength"):
@@ -1113,10 +1125,6 @@ class ExtractBase:
 
     def __init__(self):
         self.exp_type = ""
-        """
-        Issue #1781.
-        self.instrument_name = ""
-        """
         self.dispaxis = None
         self.spectral_order = None
         self.xstart = None
@@ -1132,6 +1140,7 @@ class ExtractBase:
         self.p_bkg = None
         self.smoothing_length = 0
         self.bkg_order = 0
+        self.apply_nod_offset = None
         self.nod_correction = 0.
         self.wcs = None
 
@@ -1145,156 +1154,190 @@ class ExtractBase:
 
 
     def offset_from_offset(self, input_model, slit, verbose):
-        """Get nod/dither pixel offset from [xy]_offset.
+        """Get nod/dither pixel offset from the target coordinates.
 
         Parameters
         ----------
         input_model : data model
             The input science data.
 
-        slit : SlitModel, or "dummy"
-            One slit from a MultiSlitModel (or similar), or "dummy" if
+        slit : SlitModel or None
+            One slit from a MultiSlitModel (or similar), or None if
             there are no slits.
-
-        verbose : boolean
-            If True, write log messages.
 
         Returns
         -------
         offset : float
             The offset of the exposure from the nominal position, due to
-            nod or dither.  This is the component of the offset
+            nod and/or dither.  This is the component of the offset
             perpendicular to the dispersion direction.  A positive value
             means that the spectrum is at a larger pixel number than the
             nominal location.
+
+        locn : float or None
+            The pixel coordinate of the target in the cross-dispersion
+            direction, at the middle of the spectrum in the dispersion
+            direction.
+
+        verbose : bool
+            If True, log messages.
         """
 
-        instrument = input_model.meta.instrument.name
+        # Use the WCS function to find the cross-dispersion (XD) location that
+        # is closest to the target coordinates.  This is the "actual" location
+        # of the spectrum, so the extraction region should be centered here.
+        (middle, middle_wl, locn) = self.locn_from_wcs(
+                                        input_model,
+                                        slit,
+                                        verbose)
+        if middle is not None and verbose:
+            log.debug("Spectrum location from WCS used column (or row) %d",
+                      middle)
 
-        total_points = input_model.meta.dither.total_points
-        if total_points is None or total_points < 2:
+        # Find the nominal extraction location, i.e. the XD location
+        # specified in the reference file prior to adding any nod/dither
+        # offset.  The difference is the nod/dither offset.
+        offset = 0.
+        if middle is not None and locn is not None:
+            nominal_location = self.nominal_locn(middle, middle_wl)
             if verbose:
-                log.info("Total number of dither points = %s; assuming no "
-                         "nod/dither offset", str(total_points))
-            return 0.
-
-        missing = False
-        if 'detector' not in self.wcs.available_frames:
-            if verbose:
-                log.warning("detector frame is not available,")
-            missing = True
-        if 'v2v3' not in self.wcs.available_frames:
-            if verbose:
-                if missing:
-                    log.warning("and v2v3 frame is not available,")
-                else:
-                    log.warning("v2v3 frame is not available,")
-            missing = True
-        if missing:
-            if verbose:
-                log.warning("so can't compute nod/dither offset")
-            return 0.
-        v2v3_detector = self.wcs.get_transform('v2v3', 'detector')
-
-        xoffset = input_model.meta.dither.x_offset      # in arcsec
-        yoffset = input_model.meta.dither.y_offset      # in arcsec
-        if verbose:
-            log.debug("xoffset = %s, yoffset = %s", str(xoffset), str(yoffset))
-        if xoffset is None or yoffset is None:
-            if verbose:
-                log.warning("XOFFSET and/or YOFFSET not found; "
-                            "assuming no nod/dither offset")
-            return 0.
-
-        v2ref = None
-        v3ref = None
-        v3idlyangle = None
-        vparity = None
-        wl_start = None
-        wl_end = None
-        if slit != DUMMY:
-            if hasattr(slit.meta.wcsinfo, "v2_ref"):
-                v2ref = slit.meta.wcsinfo.v2_ref                # in arcsec
-            if hasattr(slit.meta.wcsinfo, "v3_ref"):
-                v3ref = slit.meta.wcsinfo.v3_ref                # in arcsec
-            if hasattr(slit.meta.wcsinfo, "v3yangle"):
-                v3idlyangle = slit.meta.wcsinfo.v3yangle        # in degrees
-            if hasattr(slit.meta.wcsinfo, "vparity"):
-                vparity = slit.meta.wcsinfo.vparity
-            # These wavelengths are in meters; convert to microns later.
-            if hasattr(slit.meta.wcsinfo, "waverange_start"):
-                wl_start = slit.meta.wcsinfo.waverange_start
-            if hasattr(slit.meta.wcsinfo, "waverange_end"):
-                wl_end = slit.meta.wcsinfo.waverange_end
-        else:
-            if hasattr(input_model.meta.wcsinfo, "v2_ref"):
-                v2ref = input_model.meta.wcsinfo.v2_ref         # in arcsec
-            if hasattr(input_model.meta.wcsinfo, "v3_ref"):
-                v3ref = input_model.meta.wcsinfo.v3_ref         # in arcsec
-            if hasattr(input_model.meta.wcsinfo, "v3yangle"):
-                v3idlyangle = input_model.meta.wcsinfo.v3yangle # in degrees
-            if hasattr(input_model.meta.wcsinfo, "vparity"):
-                vparity = input_model.meta.wcsinfo.vparity
-            if hasattr(input_model.meta.wcsinfo, "waverange_start"):
-                wl_start = input_model.meta.wcsinfo.waverange_start
-            if hasattr(input_model.meta.wcsinfo, "waverange_end"):
-                wl_end = input_model.meta.wcsinfo.waverange_end
-
-        if v3idlyangle is None and instrument == "NIRSPEC":
-            v3idlyangle = 138.892975
-            if verbose:
-                log.warning("v3yangle not specified; "
-                            "setting it to %.6f degrees", v3idlyangle)
-        if vparity is None:
-            vparity = -1.
-            if verbose:
-                log.warning("vparity not specified; setting it to %g",
-                            vparity)
-
-        if verbose:
-            log.debug("v2ref = %s, v3ref = %s, v3yangle = %s, "
-                      "vparity = %s, wl_start = %s, wl_end = %s",
-                      str(v2ref), str(v3ref), str(v3idlyangle), str(vparity),
-                        str(wl_start), str(wl_end))
-        if (v2ref is None or v3ref is None or
-            v3idlyangle is None or vparity is None or
-            wl_start is None or wl_end is None):
+                log.debug("Target spectrum is at %g in the cross-dispersion "
+                          "direction", locn)
+            if nominal_location is not None:
                 if verbose:
-                    log.warning("Missing wcsinfo values; "
-                                "can't compute nod/dither offset")
-                return 0.
+                    log.debug("and the nominal XD location of the target "
+                              "spectrum is %g", nominal_location)
+                offset = locn - nominal_location
+            else:
+                if verbose:
+                    log.debug("but couldn't determine the nominal XD location.")
 
-        idl_v23 = trmodels.IdealToV2V3(v3idlyangle, v2ref, v3ref, vparity)
-
-        # Wavelength at the middle of the range, in microns.
-        wavelength = (0.5 * (wl_end - wl_start) + wl_start) * 1.e6
-
-        # Compute the location in V2,V3 [in arcsec]
-        xv0, yv0 = idl_v23(0., 0.)
-        xv, yv = idl_v23(xoffset, yoffset)
-        (x0, y0) = v2v3_detector(xv0, yv0, wavelength)
-        (x, y) = v2v3_detector(xv, yv, wavelength)
-        if verbose:
-            log.debug("x0 = %s, x = %s, y0 = %s, y = %s",
-                      str(x0), str(x), str(y0), str(y))
-        if x0 is None or y0 is None or x is None or y is None:
-            if verbose:
-                log.warning("One or more of x, y, x0, y0 is None; "
-                            "can't compute nod/dither offset")
-            return 0.
-        # offsets from xoffset = 0, yoffset = 0
-        dx = x - x0
-        dy = y - y0
-        if self.dispaxis == HORIZONTAL:
-            offset = dy
-        else:
-            offset = dx
         if np.isnan(offset):
             if verbose:
                 log.warning("Nod/dither offset is NaN; setting it to 0.")
             offset = 0.
+        self.nod_correction = offset
 
-        return offset
+        return offset, locn
+
+    def locn_from_wcs(self, input_model, slit, verbose):
+        """Get the location of the spectrum, based on the WCS.
+
+        Parameters
+        ----------
+        input_model : data model
+            The input science model.  This will only be used if `slit`
+            is None
+
+        slit : one slit from a MultiSlitModel (or similar), or None
+            The WCS and target coordinates will be gotten from `slit`
+            unless `slit` is None, and in that case they will be gotten
+            from `input_model`.
+
+        verbose : bool
+            If True, log messages.
+
+        Returns
+        -------
+        middle : int or None
+            Pixel coordinate in the dispersion direction within the 2-D
+            cutout (or the entire input image) at the middle of the WCS
+            bounding box.  This is the point at which to determine the
+            nominal extraction location, in case it varies along the
+            spectrum.  The offset will then be the difference between
+            `locn` (below) and the nominal location.
+
+        middle_wl : float or None
+            The wavelength at pixel `middle`.
+
+        locn : float or None
+            Pixel coordinate in the cross-dispersion direction within the
+            2-D cutout (or the entire input image) that has right ascension
+            and declination coordinates corresponding to the target location.
+            The spectral extraction region should be centered here.
+            None will be returned for `middle`, `middle_wl`, and `locn`
+            if there was not sufficient information available, e.g. if the
+            wavelength attribute or wcs function is not defined.
+        """
+
+        bb = self.wcs.bounding_box          # ((x0, x1), (y0, y1))
+
+        if bb is None:
+            if slit is None:
+                shape = input_model.data.shape
+            else:
+                shape = slit.data.shape
+            bb = wcs_bbox_from_shape(shape)
+
+        if self.dispaxis == HORIZONTAL:
+            # Width (height) in the cross-dispersion direction, from the
+            # start of the 2-D cutout (or of the full image) to the upper
+            # limit of the bounding box.  This may be smaller than the full
+            # width of the image, but it's all we need to consider.
+            xd_width = int(round(bb[1][1]))     # must be an int
+            # This is the middle of the bounding_box in the dispersion
+            # direction.
+            middle = int((bb[0][0] + bb[0][1]) / 2.)
+            x = np.empty(xd_width, dtype=np.float64)
+            x[:] = float(middle)
+            y = np.arange(xd_width, dtype=np.float64)
+            lower = bb[1][0]
+            upper = bb[1][1]
+        else:                                   # dispaxis = VERTICAL
+            xd_width = int(round(bb[0][1]))     # must be an int
+            middle = int((bb[1][0] + bb[1][1]) / 2.)
+            x = np.arange(xd_width, dtype=np.float64)
+            y = np.empty(xd_width, dtype=np.float64)
+            y[:] = float(middle)
+            lower = bb[0][0]
+            upper = bb[0][1]
+
+        # We need stuff[2], a 1-D array of wavelengths crossing the
+        # spectrum near its middle.
+        stuff = self.wcs(x, y)
+        middle_wl = np.nanmean(stuff[2])
+
+        targ_ra = input_model.meta.target.ra
+        targ_dec = input_model.meta.target.dec
+
+        x_y = self.wcs.backward_transform(targ_ra, targ_dec, middle_wl)
+
+        # locn is the xd location of the spectrum:
+        if self.dispaxis == HORIZONTAL:
+            locn = x_y[1]
+        else:
+            locn = x_y[0]
+        if locn < lower or locn > upper and targ_ra > 340.:
+            # Try this as a temporary workaround.
+            x_y = self.wcs.backward_transform(targ_ra - 360., targ_dec,
+                                              middle_wl)
+            if self.dispaxis == HORIZONTAL:
+                temp_locn = x_y[1]
+            else:
+                temp_locn = x_y[0]
+            if temp_locn >= lower and temp_locn <= upper:
+                # Subtracting 360 from the right ascension worked!
+                locn = temp_locn
+                if verbose:
+                    log.warning("targ_ra changed from %g to %g",
+                                targ_ra, targ_ra - 360.)
+
+        # If the target is at the edge of the image or at the edge of the
+        # non-NaN area, we can't use the WCS to find the location of the
+        # target spectrum.
+        if locn < lower or locn > upper:
+            if verbose:
+                log.warning("WCS implies the target is at %g, which is "
+                            "outside the bounding box, so we can't get "
+                            "target location using the WCS.", locn)
+            locn = None
+
+        return middle, middle_wl, locn
+
+    def nominal_locn(self, middle, middle_wl):
+        # Implemented in the subclasses.
+        raise NotImplementedError()
+        pass
 
 
 class ExtractModel(ExtractBase):
@@ -1310,7 +1353,8 @@ class ExtractModel(ExtractBase):
                  smoothing_length=0, bkg_order=0, nod_correction=0.,
                  x_center=None, y_center=None,
                  inner_bkg=None, outer_bkg=None, method='subpixel',
-                 subtract_background=None):
+                 subtract_background=None,
+                 apply_nod_offset=None):
         """Create a polynomial model from coefficients.
 
         Extended summary
@@ -1323,7 +1367,7 @@ class ExtractModel(ExtractBase):
         input_model : data model
             The input science data.
 
-        slit : an input slit, or a dummy value if not used
+        slit : an input slit, or None if not used
             For MultiSlit or MultiProduct data, `slit` is one slit from
             a list of slits in the input.  For other types of data, `slit`
             will not be used.
@@ -1408,6 +1452,12 @@ class ExtractModel(ExtractBase):
             If None, the value in the extract_1d reference file will be used.
             If not None, this parameter overrides the value in the
             extract_1d reference file.
+
+        apply_nod_offset : bool or None
+            If True, the target and background positions specified in the
+            reference file (or the default position, if there is no
+            reference file) will be shifted to account for nod and/or
+            dither offset.
         """
 
         super().__init__()
@@ -1483,7 +1533,7 @@ class ExtractModel(ExtractBase):
         else:
             self.extract_width = int(round(extract_width))
         # 'wavelength' or 'pixel', the independent variable for functions
-        # for lower and upper limits of source and background regions.
+        # for lower and upper limits of target and background regions.
         self.independent_var = independent_var.lower()
         if (self.independent_var != "wavelength" and
             self.independent_var != "pixel" and
@@ -1527,6 +1577,7 @@ class ExtractModel(ExtractBase):
             smoothing_length += 1               # must be odd
         self.smoothing_length = smoothing_length
         self.bkg_order = bkg_order
+        self.apply_nod_offset = apply_nod_offset
         self.nod_correction = nod_correction
 
         self.wcs = None                         # initial value
@@ -1539,13 +1590,71 @@ class ExtractModel(ExtractBase):
                     raise InvalidSpectralOrderNumberError(
                                 "Spectral order {} is not valid"
                                 .format(self.spectral_order))
-        elif slit == DUMMY:
+        elif slit is None:
             if hasattr(input_model.meta, 'wcs'):
                 self.wcs = input_model.meta.wcs
         elif hasattr(slit, 'meta') and hasattr(slit.meta, 'wcs'):
             self.wcs = slit.meta.wcs
         if self.wcs is None:
             log.warning("WCS function not found in input.")
+
+
+    def nominal_locn(self, middle, middle_wl):
+        """Find the nominal cross-dispersion location of the target spectrum.
+
+        This version is for the case that the reference file is a JSON file,
+        or that there is no reference file.
+
+        Parameters
+        ----------
+        middle: int
+            The zero-indexed pixel number of the point in the dispersion
+            direction at which `locn_from_wcs` determined the actual
+            location (in the cross-dispersion direction) of the target
+            spectrum.  This is used for evaluating the polynomial
+            functions if the independent variable is pixel.
+
+        middle_wl: float
+            The wavelength at pixel `middle`.  This is only used if the
+            independent variable for polynomial functions is wavelength.
+
+        Returns
+        -------
+        location: float or None
+            The nominal cross-dispersion location (i.e. unmodified by
+            nod or dither offset) of the target spectrum.
+        """
+
+        if self.src_coeff is None:
+            if self.dispaxis == HORIZONTAL:
+                location = float(self.ystart + self.ystop) / 2.
+            else:
+                location = float(self.xstart + self.xstop) / 2.
+        else:
+            if self.independent_var.startswith("wavelength"):
+                x = float(middle_wl)
+            else:
+                x = float(middle)
+            # Create the polynomial functions.  We'll do this again later,
+            # after adding the nod/dither offset to the coefficients, but
+            # we need to evaluate them at x now in order to get the nominal
+            # location of the spectrum.
+            self.assign_polynomial_limits(verbose=False)
+            n_srclim = len(self.p_src)
+            sum_data = 0.
+            sum_weights = 0.
+            for i in range(n_srclim):
+                lower = self.p_src[i][0](x)
+                upper = self.p_src[i][1](x)
+                weight = (upper - lower)
+                sum_data += weight * (lower + upper) / 2.
+                sum_weights += weight
+            if sum_weights == 0.:
+                location = None
+            else:
+                location = sum_data / sum_weights
+
+        return location
 
 
     def add_nod_correction(self, verbose):
@@ -1956,7 +2065,8 @@ class ImageExtractModel(ExtractBase):
                  dispaxis=HORIZONTAL,
                  smoothing_length=0,
                  nod_correction=0,
-                 subtract_background=None):
+                 subtract_background=None,
+                 apply_nod_offset=None):
         """Extract using a reference image to define the extraction and
            background regions.
 
@@ -1965,7 +2075,7 @@ class ImageExtractModel(ExtractBase):
         input_model : data model
             The input science data.
 
-        slit : an input slit, or a dummy value if not used
+        slit : an input slit, or None if not used
             For MultiSlit or MultiProduct data, `slit` is one slit from
             a list of slits in the input.  For other types of data, `slit`
             will not be used.
@@ -2003,20 +2113,21 @@ class ImageExtractModel(ExtractBase):
             will be moved to [y0 + nod, x0], where `nod` is
             int(round(nod_correction)), if the dispersion direction is
             horizontal.
+
+        apply_nod_offset : bool or None
+            If True, the reference image will be shifted by an integral
+            number of pixels to account for the nod and/or dither offset.
         """
 
         super().__init__()
 
         self.exp_type = input_model.meta.exposure.type
-        """
-        issue #1781
-        self.instrument_name = input_model.meta.instrument.name
-        """
         # ref_model contains one or more images; ref_image is the one that
         # matches the current configuration (slit name and spectral order).
         self.ref_image = ref_image
         self.spectral_order = spectral_order
         self.dispaxis = dispaxis
+        self.apply_nod_offset = apply_nod_offset
         self.nod_correction = nod_correction
 
         if smoothing_length is None:
@@ -2038,13 +2149,72 @@ class ImageExtractModel(ExtractBase):
                     raise InvalidSpectralOrderNumberError(
                                 "Spectral order {} is not valid"
                                 .format(self.spectral_order))
-        elif slit == DUMMY:
+        elif slit is None:
             if hasattr(input_model.meta, 'wcs'):
                 self.wcs = input_model.meta.wcs
         elif hasattr(slit, 'meta') and hasattr(slit.meta, 'wcs'):
             self.wcs = slit.meta.wcs
         if self.wcs is None:
             log.warning("WCS function not found in input.")
+
+
+    def nominal_locn(self, middle, middle_wl):
+        """Find the nominal cross-dispersion location of the target spectrum.
+
+        This version is for the case that the reference file is an image.
+
+        Parameters
+        ----------
+        middle: int
+            The zero-indexed pixel number of the point in the dispersion
+            direction at which `locn_from_wcs` determined the actual
+            location (in the cross-dispersion direction) of the target
+            spectrum.
+
+        middle_wl: float
+            The wavelength at pixel `middle`.  This is not used in this
+            version.
+
+        Returns
+        -------
+        location: float or None
+            The nominal cross-dispersion location (i.e. unmodified by
+            nod or dither offset) of the target spectrum.
+            The value will be None if `middle` is outside the reference
+            image or if the reference image does not specify any pixels
+            to extract at `middle`.
+        """
+
+        shape = self.ref_image.data.shape
+
+        bad = False
+        if self.dispaxis == HORIZONTAL:
+            if middle >= 0 and middle < shape[1]:
+                middle_line = self.ref_image.data[:, middle]
+            else:
+                bad = True
+        else:
+            if middle >= 0 and middle < shape[0]:
+                middle_line = self.ref_image.data[middle, :]
+            else:
+                bad = True
+        if bad:
+            log.warning("Can't determine nominal location of target "
+                        "spectrum because middle = %g is off the image.",
+                        middle)
+            return None
+
+        mask_target = np.where(middle_line > 0., 1., 0.)
+        x = np.arange(len(middle_line), dtype=np.float64)
+
+        numerator = (x * mask_target).sum()
+        denominator = mask_target.sum()
+        if denominator > 0.:
+            location = numerator / denominator
+        else:
+            location = None
+
+        return location
 
 
     def add_nod_correction(self, verbose):
@@ -2290,14 +2460,6 @@ class ImageExtractModel(ExtractBase):
                     dec[:] = -999.
                     wcs_wl[:] = -999.
             else:
-                """
-                See issue #1781
-                if self.instrument_name == "NIRSPEC":
-                    # xxx temporary:  NIRSpec wcs is one-based.
-                    ra, dec, wcs_wl = self.wcs(x_array + 1., y_array + 1.)
-                else:
-                    ra, dec, wcs_wl = self.wcs(x_array, y_array)
-                """
                 ra, dec, wcs_wl = self.wcs(x_array, y_array)
             # We need one right ascension and one declination, representing
             # the direction of pointing.
@@ -2400,7 +2562,7 @@ class ImageExtractModel(ExtractBase):
 
 
     def separate_target_and_background(self, ref):
-        """Create masks for source and background.
+        """Create masks for target and background.
 
         Parameters
         ----------
@@ -2509,7 +2671,7 @@ def interpolate_response(wavelength, relsens, verbose):
 
 
 def run_extract1d(input_model, refname, smoothing_length, bkg_order,
-                  log_increment, subtract_background):
+                  log_increment, subtract_background, apply_nod_offset):
     """Extract 1-D spectra.
 
     This just reads the reference file (if any) and calls do_extract1d.
@@ -2522,10 +2684,10 @@ def run_extract1d(input_model, refname, smoothing_length, bkg_order,
     refname : str
         The name of the reference file, or "N/A".
 
-    smoothing_length : int
+    smoothing_length : int or None
         Width of a boxcar function for smoothing the background regions.
 
-    bkg_order : int
+    bkg_order : int or None
         Polynomial order for fitting to each column (or row, if the
         dispersion is vertical) of background.
 
@@ -2540,6 +2702,11 @@ def run_extract1d(input_model, refname, smoothing_length, bkg_order,
         If None, the value in the extract_1d reference file will be used.
         If not None, this parameter overrides the value in the
         extract_1d reference file.
+
+    apply_nod_offset : bool or None
+        If True, the target and background positions specified in the
+        reference file (or the default position, if there is no reference
+        file) will be shifted to account for nod and/or dither offset.
 
     Returns
     -------
@@ -2560,7 +2727,8 @@ def run_extract1d(input_model, refname, smoothing_length, bkg_order,
 
     output_model = do_extract1d(input_model, ref_dict,
                                 smoothing_length, bkg_order,
-                                log_increment, subtract_background)
+                                log_increment, subtract_background,
+                                apply_nod_offset)
 
     return output_model
 
@@ -2600,8 +2768,9 @@ def ref_dict_sanity_check(ref_dict):
     return ref_dict
 
 
-def do_extract1d(input_model, ref_dict, smoothing_length, bkg_order,
-                 log_increment, subtract_background):
+def do_extract1d(input_model, ref_dict, smoothing_length=None,
+                 bkg_order=None, log_increment=50,
+                 subtract_background=None, apply_nod_offset=None):
     """Extract 1-D spectra.
 
     Parameters
@@ -2613,10 +2782,10 @@ def do_extract1d(input_model, ref_dict, smoothing_length, bkg_order,
         The contents of the reference file.  This will be None if there
         is no reference file (i.e. if refname was "N/A").
 
-    smoothing_length : int
+    smoothing_length : int or None
         Width of a boxcar function for smoothing the background regions.
 
-    bkg_order : int
+    bkg_order : int or None
         Polynomial order for fitting to each column (or row, if the
         dispersion is vertical) of background.
 
@@ -2626,10 +2795,16 @@ def do_extract1d(input_model, ref_dict, smoothing_length, bkg_order,
         `log_increment` integrations.
 
     subtract_background : bool or None
-        User supplied flag indicating whether the background should be subtracted.
+        User supplied flag indicating whether the background should be
+        subtracted.
         If None, the value in the extract_1d reference file will be used.
         If not None, this parameter overrides the value in the
         extract_1d reference file.
+
+    apply_nod_offset : bool or None
+        If True, the target and background positions specified in the
+        reference file (or the default position, if there is no reference
+        file) will be shifted to account for nod and/or dither offset.
 
     Returns
     -------
@@ -2650,6 +2825,22 @@ def do_extract1d(input_model, ref_dict, smoothing_length, bkg_order,
     # This will be relevant if we're asked to extract a spectrum and the
     # spectral order is zero.  That's only OK if the disperser is a prism.
     prism_mode = is_prism(input_model)
+
+    if apply_nod_offset:
+        exp_type = input_model.meta.exposure.type.upper()
+        source_type = input_model.meta.target.source_type.upper()
+        if exp_type in WFSS_EXPTYPES + ['NRS_FIXEDSLIT', 'NRS_MSASPEC']:
+            apply_nod_offset = False
+            log.warning("Correcting for nod/dither offset is currently "
+                        "not supported for exp_type = %s, so "
+                        "apply_nod_offset will be set to False",
+                        input_model.meta.exposure.type)
+        if source_type != 'POINT':
+            apply_nod_offset = False
+            log.warning("SRCTYPE = '%s'; correcting for nod/dither "
+                        "offset will only be done for a point source, ",
+                        "so apply_nod_offset will be set to False",
+                        input_model.meta.target.source_type)
 
     if isinstance(input_model, datamodels.MultiSlitModel) or \
        isinstance(input_model, datamodels.MultiProductModel):
@@ -2673,7 +2864,8 @@ def do_extract1d(input_model, ref_dict, smoothing_length, bkg_order,
             extract_params = get_extract_parameters(
                                 ref_dict,
                                 slit, slit.name, sp_order,
-                                input_model.meta, smoothing_length, bkg_order)
+                                input_model.meta, smoothing_length, bkg_order,
+                                apply_nod_offset)
             if subtract_background is not None:
                 extract_params['subtract_background'] = subtract_background
             if extract_params['match'] == NO_MATCH:
@@ -2764,11 +2956,12 @@ def do_extract1d(input_model, ref_dict, smoothing_length, bkg_order,
                                     ref_dict,
                                     input_model, slitname, sp_order,
                                     input_model.meta, smoothing_length,
-                                    bkg_order)
+                                    bkg_order,
+                                    apply_nod_offset)
                 if subtract_background is not None:
                     extract_params['subtract_background'] = subtract_background
                 if extract_params['match'] == EXACT:
-                    slit = DUMMY
+                    slit = None
                     find_dispaxis(input_model, slit, sp_order, extract_params)
                     if extract_params['dispaxis'] is None:
                         log.warning("The dispersion direction couldn't be "
@@ -2831,7 +3024,7 @@ def do_extract1d(input_model, ref_dict, smoothing_length, bkg_order,
         elif isinstance(input_model, (datamodels.CubeModel,
                                       datamodels.SlitModel)):
 
-            slit = DUMMY
+            slit = None
 
             # NRS_BRIGHTOBJ exposures are instances of SlitModel.
             prev_offset = OFFSET_NOT_ASSIGNED_YET
@@ -2847,7 +3040,8 @@ def do_extract1d(input_model, ref_dict, smoothing_length, bkg_order,
                                     ref_dict,
                                     input_model, slitname, sp_order,
                                     input_model.meta, smoothing_length,
-                                    bkg_order)
+                                    bkg_order,
+                                    apply_nod_offset)
                 if subtract_background is not None:
                     extract_params['subtract_background'] = subtract_background
                 if extract_params['match'] == NO_MATCH:
@@ -2952,7 +3146,8 @@ def do_extract1d(input_model, ref_dict, smoothing_length, bkg_order,
             except AttributeError:
                 source_type = "unknown"
             output_model = ifu.ifu_extract1d(input_model, ref_dict,
-                                             source_type, subtract_background)
+                                             source_type, subtract_background,
+                                             apply_nod_offset)
 
         else:
             log.error("The input file is not supported for this step.")
@@ -3247,8 +3442,8 @@ def extract_one_slit(input_model, slit, integ,
     input_model : data model
         The input science model.
 
-    slit : one slit from a MultiSlitModel (or similar), or "dummy"
-        If slit is "dummy", the data array is input_model.data; otherwise,
+    slit : one slit from a MultiSlitModel (or similar), or None
+        If slit is None, the data array is input_model.data; otherwise,
         the data array is slit.data.
         In the former case, if `integ` is zero or larger, the spectrum
         will be extracted from the 2-D slice input_model.data[integ].
@@ -3311,7 +3506,7 @@ def extract_one_slit(input_model, slit, integ,
             wl_array = input_model.wavelength
         except AttributeError:
             wl_array = None
-    elif slit == DUMMY:
+    elif slit is None:
         data = input_model.data
         if hasattr(input_model, 'dq'):
             input_dq = input_model.dq
@@ -3344,12 +3539,17 @@ def extract_one_slit(input_model, slit, integ,
 
     # Only call this method for the first integration.
     if prev_offset == OFFSET_NOT_ASSIGNED_YET:
-        offset = extract_model.offset_from_offset(input_model, slit, verbose)
-        if offset != 0:                         # xxx should be temporary
-            if verbose:
-                log.debug("Computed nod/dither offset = %s, but don't "
-                          "trust this yet, so assuming 0", str(offset))
-            offset = 0.                         # xxx should be temporary
+        # If apply_nod_offset is False, compute the offset anyway in order
+        # to log the offset and target location (below), but don't log any
+        # messages in or below offset_from_offset.
+        local_verbose = extract_model.apply_nod_offset
+        (offset, locn) = extract_model.offset_from_offset(
+                                input_model, slit, local_verbose)
+        if verbose:
+            log.debug("Computed nod/dither offset = %s, "
+                      "target location = %s.", str(offset), str(locn))
+        if not extract_model.apply_nod_offset:
+            offset = 0.
     else:
         offset = prev_offset
     extract_model.nod_correction = offset