+++ jwst/extract_1d/extract_1d_step.py
@@ -43,6 +43,21 @@ class Extract1dStep(Step):
         If None, the value in the extract_1d reference file will be used.
         If not None, this parameter overrides the value in the
         extract_1d reference file.
+
+    apply_nod_offset : bool or None
+        If True, the source and background positions specified in the
+        reference file (or the default position, if there is no reference
+        file) will be shifted to account for nod and/or dither offset.  If
+        None (the default), the value in the reference file will be used,
+        or it will be set to True if it is not specified in the reference
+        file.  This offset is determined by finding the location in the data
+        corresponding to the target position (keywords TARG_RA and TARG_DEC).
+        For NIRSpec fixed-slit or MOS data, there must be different target
+        coordinates for each slit for which a nod correction might be
+        needed, and this is not implemented yet in extract_1d.
+        It also doesn't make sense to apply a nod/dither offset for an
+        extended target, so this flag can internally be overridden (set to
+        False) for extended targets.
     """
 
     spec = """
@@ -55,6 +70,11 @@ class Extract1dStep(Step):
     log_increment = integer(default=50)
     # Flag indicating whether the background should be subtracted.
     subtract_background = boolean(default=None)
+    # If True, the locations of the target and background regions will be
+    # shifted to correct for the computed nod/dither offset.
+    # Currently this offset is not applied for NIRSpec fixed-slit or
+    # MOS (MSA) data), or for WFSS data.
+    apply_nod_offset = boolean(default=None)
     """
 
     reference_file_types = ['extract1d']
@@ -123,7 +143,8 @@ class Extract1dStep(Step):
                                                  self.smoothing_length,
                                                  self.bkg_order,
                                                  self.log_increment,
-                                                 self.subtract_background)
+                                                 self.subtract_background,
+                                                 self.apply_nod_offset)
                     # Set the step flag to complete in each MultiSpecModel
                     temp.meta.cal_step.extract_1d = 'COMPLETE'
                     result.append(temp)
@@ -142,7 +163,8 @@ class Extract1dStep(Step):
                                                self.smoothing_length,
                                                self.bkg_order,
                                                self.log_increment,
-                                               self.subtract_background)
+                                               self.subtract_background,
+                                               self.apply_nod_offset)
                 # Set the step flag to complete
                 result.meta.cal_step.extract_1d = 'COMPLETE'
             else:
@@ -163,7 +185,8 @@ class Extract1dStep(Step):
                                            self.smoothing_length,
                                            self.bkg_order,
                                            self.log_increment,
-                                           self.subtract_background)
+                                           self.subtract_background,
+                                           self.apply_nod_offset)
             # Set the step flag to complete
             result.meta.cal_step.extract_1d = 'COMPLETE'